+++ datasource_sse.go
@@ -153,7 +153,7 @@ func (ds *SseDataSource) loadData(ctx context.Context) error {
 		return nil
 	}
 
-	err = ds.client.UpdateFromApiResponse(resp)
+	err = ds.client.updateFromApiResponse(ctx, resp)
 	if err != nil {
 		return err
 	}